+++ src/genie/core/parameter/descriptor.cc
@@ -19,7 +19,7 @@ bool Descriptor::equals(const Descriptor *desc) const { return dec_cfg_preset ==
 
 // ---------------------------------------------------------------------------------------------------------------------
 
-void Descriptor::write(util::BitWriter &writer) const { writer.writeBits(static_cast<uint8_t>(dec_cfg_preset), 8); }
+void Descriptor::write(util::BitWriter &writer) const { writer.writeBits(dec_cfg_preset, 8); }
 std::unique_ptr<Descriptor> Descriptor::factory(GenDesc desc, util::BitReader &reader) {
     auto preset = reader.read<uint8_t>();
     UTILS_DIE_IF(preset != 0, "Invalid DecCfgPreset");
