+++ src/genie/entropy/paramcabac/decoder.cc
@@ -39,13 +39,13 @@ DecoderRegular::DecoderRegular(core::GenDesc desc, util::BitReader &reader)
 // ---------------------------------------------------------------------------------------------------------------------
 
 void DecoderRegular::setSubsequenceCfg(uint8_t index, Subsequence &&cfg) {
-    descriptor_subsequence_cfgs[static_cast<uint8_t>(index)] = cfg;
+    descriptor_subsequence_cfgs[index] = cfg;
 }
 
 // ---------------------------------------------------------------------------------------------------------------------
 
 const Subsequence &DecoderRegular::getSubsequenceCfg(uint8_t index) const {
-    return descriptor_subsequence_cfgs[static_cast<uint8_t>(index)];
+    return descriptor_subsequence_cfgs[index];
 }
 
 // ---------------------------------------------------------------------------------------------------------------------
@@ -105,13 +105,13 @@ DecoderTokenType::DecoderTokenType(core::GenDesc desc, util::BitReader &reader)
 // ---------------------------------------------------------------------------------------------------------------------
 
 void DecoderTokenType::setSubsequenceCfg(uint8_t index, Subsequence &&cfg) {
-    descriptor_subsequence_cfgs[static_cast<uint8_t>(index)] = cfg;
+    descriptor_subsequence_cfgs[index] = cfg;
 }
 
 // ---------------------------------------------------------------------------------------------------------------------
 
 const Subsequence &DecoderTokenType::getSubsequenceCfg(uint8_t index) const {
-    return descriptor_subsequence_cfgs[static_cast<uint8_t>(index)];
+    return descriptor_subsequence_cfgs[index];
 }
 
 // ---------------------------------------------------------------------------------------------------------------------
